frogger v1.0
Frogger v1.0 is the first version I would be comfortable turning in. It
technically meets the requirements of the assignment. However, if you're
looking for a fun game.. still pretty bad. It makes you wait at the
beginning of the game for a while, hitboxes are close but not exact, and
the game goes from too easy to too hard in about 3 levels.